Safari-611.1.10/LayoutTests/accessibility/img-alt-tag-only-whitespace-expected.txt




This tests that images with alt tags that only have white space are ignored.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS imagesGroup.childrenCount is 1
PASS platformValueForW3CName(imagesGroup.childAtIndex(0)) is "   Image "
PASS successfullyParsed is true

TEST COMPLETE
